feat: import all colours from figma file not just ones starting with `custom/`
