+++ package.json
@@ -83,6 +83,7 @@
     "@storybook/manager-webpack5": "^6.5.9",
     "@storybook/react": "^6.5.9",
     "@storybook/theming": "^6.5.9",
+    "@types/flat": "^5.0.2",
     "@types/jest": "^28.1.1",
     "@types/lodash.merge": "^4.6.7",
     "@types/prompts": "^2.0.14",
@@ -103,6 +104,7 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.30.0",
     "eslint-plugin-react-hooks": "^4.5.0",
+    "flat": "^5.0.2",
     "framer-motion": "^6.3.11",
     "husky": "^8.0.1",
     "jest": "^28.1.1",
